refactor: simplify phase tracking + extract useTick
- Split useAgentPhase into a tracker (useTrackAgentPhase, called once
  from RunScreen) and a reader (useAgentPhase, used by VisualizerTab
  and PhaseVisual). Store is now the single source of truth — the
  lastDetectedRef + mirror-via-useEffect dance is gone. Eliminates
  the duplicate classification work that used to run on every render.
- useTick moved to src/ui/tui/hooks/useTick.ts (used by all 6 visuals).
- Stripped narrative WHAT comments around the store atom, EQ tick, and
  ESM require fallback; kept the audio voice comments (load-bearing).
- Hoisted the EQ glyph string to module scope so it doesn't allocate
  every frame, and cached Date.now() once per VisualizerTab render.
