+++ src/ui/tui/screens/RunScreen.tsx
@@ -24,7 +24,10 @@ import { LearnCard } from '@ui/tui/components/LearnCard';
 import { TipsCard } from '@ui/tui/components/TipsCard';
 import { useStdoutDimensions } from '@ui/tui/hooks/useStdoutDimensions';
 import { useFileWatcher } from '@ui/tui/hooks/file-watcher';
-import { VisualizerTab, useAgentPhase } from '@ui/tui/components/PhaseVisuals';
+import {
+  VisualizerTab,
+  useTrackAgentPhase,
+} from '@ui/tui/components/PhaseVisuals';
 import { EVENT_PLAN_FILE } from '@lib/programs/posthog-integration/index';
 import { getProgramConfig } from '@lib/programs/program-registry';
 import { getContentBlocks as getSkillContentBlocks } from '@lib/programs/agent-skill/content/index';
@@ -55,10 +58,9 @@ export const RunScreen = ({ store }: RunScreenProps) => {
 
   const [columns] = useStdoutDimensions();
 
-  // Track the active agent phase into the store, even when the user is on
-  // a tab that doesn't render PhaseVisual — so the Visualizer's elapsed
-  // clock reflects real stage time when they switch over.
-  useAgentPhase(store);
+  // Mirror the active phase into the store regardless of which tab is open,
+  // so the Visualizer's elapsed clock reflects real stage time.
+  useTrackAgentPhase(store);
 
   const progressItems: ProgressItem[] = store.tasks.map((t) => ({
     label: t.label,
