+++ src/ui/tui/store.ts
@@ -76,11 +76,6 @@ export class WizardStore {
   private $learnCardBlockIdx = atom(0);
   private $learnCardComplete = atom(false);
   private $version = atom(0);
-  // Global "what's playing right now" track for the Visualizer tab. Set by
-  // whichever component derives the current agent phase; survives tab
-  // switches so the elapsed clock measures real stage time, not "time since
-  // the Visualizer was last mounted". Generic `stage: string` so the store
-  // stays free of product knowledge.
   private $currentStage = atom<{ stage: string; startedAt: number } | null>(
     null,
   );
@@ -243,11 +238,8 @@ export class WizardStore {
     return this.$currentStage.get();
   }
 
-  /**
-   * Record the current stage of work. If the stage hasn't changed, the
-   * stored `startedAt` is preserved so elapsed-time readouts keep counting
-   * up across re-renders and tab switches.
-   */
+  /** No-op when the stage hasn't changed, so `startedAt` survives across
+   *  re-renders and tab switches and measures real stage time. */
   setCurrentStage(stage: string): void {
     const cur = this.$currentStage.get();
     if (cur?.stage === stage) return;
